get rid of Robot `static` field
This field was used to determine whether a robot could move (to
e.g. prevent the base from moving).  Now it is governed by whether the
robot has treads or not.  (And besides, the `direction` of the base
is (0,0) so it wouldn't go anywhere even if it did have treads.)