bug fix: step each robot separately
Previously, we got the entire robot map out of the game state, updated
it all at once using the very cool `M.traverseMaybeWithKey` function,
then put the resulting updated robot map back into the game state.
Unfortunately, once we add commands such as `give` that allow one
robot to have some kind of effect on another, this no longer works:
all the updates to the `robotMap` in the game state are lost when we
replace it wholesale at the end.  Instead, we can take out each robot,
step it, and replace it one by one.