Add Hindley-Milner style polymorphism to the type system, with complete type reconstruction (#1)
- Add Hindley-Milner style rank-1 polymorphism

- remove distinction between Term, ATerm, UTerm

I wasn't really using the type annotations, and it was going to get in
the way when I start implementing Hindley-Milner.  If I really need
type annotations on typechecked terms later I'll need to do it in a more
principled way, with terms represented by a two-level type and
inserting type annotations at every level.