overhaul handling of frames and ticks
Now do a variable number of ticks per frame, depending on how long it
took to draw the previous frame.  This means we always maintain the
target ticks/second rate, regardless of the actual frame rate.

Logic taken from https://gafferongames.com/post/fix_your_timestep/ .

Still to do is to somehow deal with the "spiral of death" mentioned in
that post, which I have in fact observed when pushing the tick/s rate
high enough.