refactor: get rid of `newRobots`
The only reason we had this temporary list in which to put newly built
robots was because of the way we used to update the entire `robotMap`
all at once and then overwrite it (which would have clobbered any
robots that got added to it in the meantime).  But that was always a
stupid way to do it, and we don't do it that way any more.  Now the
`addRobot` function just directly adds the robot to the `robotMap`.