line-wrap recipes
This is a partial solution to #48, and way better than nothing.
However, it's a bit difficult to read the wrapped recipes; ideally the
continuation lines would be indented, so it would be easy to tell
visually where a new recipe starts. But it doesn't seem possible
to do this given the current `brick` API.  I've created an upstream
issue for it here: https://github.com/jtdaugherty/word-wrap/issues/8