tmQ quasiquoter, now with bonus type checking!
The `tmQ` quasiquoter now runs the entire processing pipeline on a
term instead of only parsing.  This is 100% better since now the
typechecking will fail at compile time instead of runtime.  I'm very
glad I took the time to figure out how to make this work.  In the end
it wasn't that difficult, but it did require making a bunch more
`Data` instances for things, and adding cases for `TAntiString` in
more places (e.g. the typechecker and capability checker), since now
an antiquoted string variable has to go all the way through the
pipeline before finally getting a Haskell variable substituted for it
when the resulting `ProcessedTerm` is lifted into an AST.