Add noop keyword (#138)
Adds noop keyword to the player-facing language. Previous syntax
(`{}`) is retained.  Only parsing of the language is affected, the
representation remains the same regardless of which syntax was used.
The new keyword is now the canonical version of the command, used in
pretty-printing and the like.

Fixes #101

I'm not certain about which syntax to use for the syntax info.  I decided to
go with the new `noop` keyword instead of the old `{}` because it is more
explicit and clearer.