select logger and scroll to the end when a log message is added
This way the user will be able to see right away why e.g. the command
they typed at the REPL did not work.  It also means that if a robot
crashes and is later 'view'ed, the log message will automatically be
shown first.

Also, we now distinguish between inventory and installed item entries
in the robot inventory list. This makes several things easier, and
also fixed a bug where the focus could unexpectedly jump from an
installed device to an inventory entity of the same name when the
inventory was updated.