cap maximum ticks per frame at 30
Cap the maximum number of ticks per frame (arbitrarily) at 30, to
prevent the "spiral of death" (see
https://gafferongames.com/post/fix_your_timestep/).

To see this in effect, go into Creative mode, crank up the ticks per
second to 32 or so, then (for suitable definitions of `repeat` and
`forever`) execute

`repeat 1000 { build "" {forever {move; turn left}}}`

Without this PR, it should eventually grind to a halt.  With this PR,
the game should keep running, though with lots of lag.