bug fix: initialize store when making item from inventory panel
We used to call `initMachine` with `emptyStore` but this was wrong.
It meant that the store was reset every time you hit Enter on an
inventory item to `make` it.  So e.g. if you had made a `def` and
executed it at least once, then made something from the inventory,
trying to execute the same definition again would yield a fatal
"reference to unknown memory cell" error.