fix for inventory view when uploading scanned data
The base's inventory was not updating when a robot uploaded some scan
data to it.  The problem was that after switching to homomorphic
hashing for inventories (https://github.com/byorgey/swarm/issues/229),
adding 0 copies of something to the inventory, although it still
worked properly, no longer caused the inventory hash to be updated,
and the UI uses the hash to decide when to redraw.

The solution was to let each entity with k copies in the inventory
contribute (k+1) times its hash to the inventory hash.  This way the
hash distinguishes between having 0 copies of an entity and not knowing
about it at all.