Pop up entity description on `Enter` (#335)
Hitting `Enter` on an inventory item pops up a dialog box with its description.  For items with longer descriptions this is a convenient way to be able to see the whole description without having to scroll the info box in the lower left.

`Enter` used to try to `make` the focused item; that is now accomplished with `m`.