Add recipe for `mirror` and make it required for `self` (#416)
The `whoami` command used to be more interesting back when robot names were used to uniquely identify them.  It seemed fitting to make `mirror` required for `self`, which now plays the same role.  I also added a recipe for `mirror` which we lacked (`glass` + `silver`) and made a new type of `deep mine` where we can mine `gold`, `silver`, and `mithril`.