add drilled mines to known entities
Previously, after drilling a mine, the robot did not "know about" it;
it was necessary to go send another robot to scan it.  This seemed a
little silly; if you just drilled and found a copper mine, wouldn't
you know what it is?

Specifically, when completing a drilling recipe, any non-portable
outputs of the recipe are now inserted into the robot's inventory with
count 0.