Add branch filter to reduce CI cost
The action presently runs twice because both `on` condition matches
every contribution. With this change, the push action only triggers
for the final push to the main branch, and the pull_requestion action
only triggers for PR made for the main branch.

That also means restyled PR should not trigger CI as they are made
for non main branches.