Add new `harvest` command + `harvester` device (#503)
Make a new `harvester` device + `harvest` command.  The recipe for a `harvester` is 2 `board` + 2 `wooden gear` + 1 `box`, hence it can be made completely out of 3 `tree`s. (3 trees make 12 `board`s; 6 `board`s are for the `box`, 4 are to make the 2 `wooden gear`s).

The `harvester` device provides the `harvest` and `grab` capabilities, hence it supports both the `harvest` and `grab` commands. 
The `harvest` command now does exactly what the `grab` command used to do; the `grab` command works identically except that it never leaves behind a growing seed. 

As an aside, I am actually not sure what happens if you try to `build` a robot that uses both `grab` and `harvest` (but none of the other capabilities provided by a `grabber`).  Is the capability analysis smart enough to realize that it should just install a `harvester`, or might it end up installing both a `harvester` and a `grabber` unnecessarily?

Closes #490 (I split out another part of that issue into #502).  @xsebek , I know in #490 you expressed an objection to adding a `harvester` device so I'm curious what you think about this.