don't tab complete when there are no identifier chars typed (#543)
When the prompt was blank, it didn't tab complete; however, when there
were any characters typed at all, it found the "last word" (all
identifier characters at the end) and tried to match on it.  However,
if there were no identifier chars, the resulting empty "last word"
matched everything.

This fix now has a special case to check whether the `lastWord` is
blank, which also takes care of the original special case when the
prompt is completely empty.

Fixes #518.