disallow problematic constants in `atomic`
Only allowed whitelisted constants inside `atomic`.  The idea is to
exclude constants that might take multiple ticks to execute.  It's a
whitelist on purpose (even though expressing it that way is longer) so
that we can't accidentally introduce a new problematic constant in the
future and forget to exclude it.