bug fix: don't `continueWithoutRedraw` from modal scroll handler (#607)
The problem is that we pass along events to *both* the dialog (for
e.g. switching between dialog buttons) *and* to the viewport (for e.g.
scrolling up and down).  If we hit Tab the dialog handles it, but then
the viewport handler used to say `continueWithoutRedraw` so the screen
didn't update to reflect the moved button focus.