add recipe for turning iron into a lodestone (#646)
If you rub some iron with a lodestone for a long time, the iron turns into another lodestone.

While playing the game it felt like lodestones were too scarce, given the fact that they are needed to evaluate `not`, which ends up getting used frequently.  This way lodestones are actually renewable.