WIP: Infinite inventory
This approach is buggy in two ways:
- it can crash because Natural is too strict
- it breaks the fragile hashing laws for inventory

A better approach would be to either do without
a Num instance and handle every usage carefully
or add negative infinity too and use integers.

In that case we might as well add them to the number
values in game, though that increases the risk
of the infinities being added together.
But that can be handled with a safeAdd function.