Disambiguate choice of base robot
When none of the robot definitions are assigned explicit coordinates
(i.e. all of the robots are specified via the map and palette), this
allows scenario authors to determine the "base" robot via the order
of definitions as they exist in the scenario file.

A section detailing the base selection logic has also been added
to the Scenario Authoring Guide.

closes #790