Disambiguate choice of base robot (#805)
When none of the robot definitions are assigned explicit coordinates (i.e. all of the robots are specified via the map and palette), this allows scenario authors to determine the "base" robot via the order of definitions as they exist in the scenario file.

A comment detailing the base selection logic has also been added.

closes #790