New ranching scenario (#835)
Demo with:

    stack run -- --scenario Challenges/Ranching/gated-paddock --autoplay

The primary element of this scenario is "testing for a closed loop" in the objective condition.  As written, this test is performed every "tick", and introduces noticeable rendering lag at higher "ticks / s" speeds.

What may be better is for the player to "ask" for the condition to be checked, perhaps by standing somewhere specific, picking up a particular object, or even with `say "Ready!"`.