A "bucket brigade" of robots (#885)
Demo with:

    stack run -- --scenario scenarios/Challenges/bucket-brigade.yaml --autoplay

The idea of this scenario will be to form a "bucket brigade", where, in lieu of mobility, robots form a chain to pass items along back to the base.

There needs to exist one set of "treads" so that the robots may distribute themselves into a line.  In this scenario, a single `treads` is shared and passed between each of the robots.

However, once the robots have arranged themselves, all of the robots still have "treads" equipped.
So to contrive a circumstance where the "bucket brigade" is necessary, we need some way to inhibit the robots' movement (see #883).