Parameterize scenario datatypes (#903)
Towards #873

## Motivation

The `Entity` datatype has a fair amount of state and extra fields that are not relevant to serializing a scenario (in particular, a world map) to a file.

In this PR, the `Cell` record is renamed to `PCell` (representing "Parameterized Cell"), replacing `Entity` with a type parameter.  This permits the introduction of a trimmed-down `Entity` type for the purpose of serialization.  This lightweight type can also be a reliable `Map` key.

A type alias `Cell` is created with the original `Entity` type as a parameter.

The `WorldDescription` and `WorldPalette` are likewise parameterized on the `Cell` type.