Fix capability checking, and refactor/add lots of comments (#959)
Fixes #397.  The only way I could understand this in order to fix it was to totally refactor the code and add lots of comments as I went.  I feel like this is some of the most difficult code to wrap one's head around in the codebase.  Hopefully now it's a bit easier to understand (though still not easy, I imagine).