Fix `EntityMap` merging (#962)
Fixes #961.  Merging the `entitiesByCap` maps simply with `(<>)` (which is the same as `M.union`) was not correct: map union is left-biased, so if both maps had entities keyed to the same capability, the entities from the right-hand map would simply be thrown away.  Instead, we must merge them with `M.unionWith (<>)` which will append the two `Entity` lists for a given capability.