Add `isempty : cmd bool` to check whether current cell has an entity (#968)
As I've been playing around in classic mode I've found it a bit annoying/unbalanced that although you can use `ishere : text -> cmd bool` to check for the presence of a *specific* entity, the only way to check whether a cell is empty is to write something like
```
def is_empty : cmd bool =
  s <- scan down;
  return (s == inl ())
end
```
However, this requires having a `comparator` and an `ADT calculator` --- the latter in particular is quite difficult to construct.  So, I propose adding a primitive `isempty : cmd bool`.  Note this doesn't make `scan` useless --- it's more powerful still since it lets you find out the name of anything (even stuff not in the current cell).  

An alternative might be to get rid of `ishere` as well (since it can also be programmed in terms of `scan`), but tons of obvious early-game automation (*e.g.* harvesting a tree plantation) depends on being able to have conditionals based on the presence or absence of certain entities, and it would be unreasonable to have that depend on making an `ADT calculator`.

The `scanner` now provides quite a few commands; I think that's OK but chime in if you think we ought to split it into multiple devices.  I think it makes sense to have `scan`, `ishere`, and `isempty` all provided by `scanner` since it's just providing several interfaces to the same "service", one general and several simpler but easier-to-use versions.  If we split anything out it could make sense to split out `blocked` (`laser range finder`, perhaps?) and/or `upload`.