Todo enforcement line number (#1045)
As an extension to #514, include the line number in the grep output, and move the grep logic to a script, so that it can be used within VS Code and clicking on the file:line will open the editor at the right place.

![vscode-demo](https://user-images.githubusercontent.com/261693/214456476-c27b2360-1018-4281-8930-45faaa17d66b.png)

The GitHub action still works.  See this CI run: https://github.com/swarm-game/swarm/actions/runs/4001861308/jobs/6868527786