bind tutorial: give base a `compass` and equip it with a `logger`
There's no reason not to allow a `compass` in this challenge, if
people had become comfortable typing things like `turn north`.  Later
we will have to break it to them that they will have to do without
this for a while...

Also, the base absolutely ought to have a `logger` equipped.  As it
is, if the user tries to `build` a robot that does something the base
does not have the devices for, it simply fails silently.