add random "static" to `view` outside a certain range (#1110)
The basic principle here is that you cannot communicate with robots that are too far away.  We also want to uphold the principle that it is not possible to instantly communicate information over arbitrary distances.

Concretely, there is now an "inner radius" (16 by default) and an "outer radius" (64 by default).
* Inside of the inner radius, communication is perfect.
* Between the inner and outer radius, the `view` becomes increasingly covered with "static".
* Outside the outer radius, the view is completely static, *and* it is impossible to see a viewed robot's inventory.

In particular there should be absolutely no difference between trying to `view` a robot that is deceased vs a robot that is too far away.  A nice bonus feature is that if you are `view`ing a robot that is suddenly destroyed, e.g. by driving into the water, the entire view instantly turns to static, which looks cool.

The static is based on a hash of the offset from the view center to the cell + the current tick count / 16.  That means that the static pattern (1) doesn't shift when the viewed robot moves around, and (2) randomly changes every second or so.

[See this asciinema cast](https://asciinema.org/a/564988) for a demo of it in action.

Closes #17 .

- [x] update `antenna` description