get rid of ill-conceived `export` primitive
It's rarely useful in practice to export *everything* in scope.  At
the REPL, it includes all the `itN` variables.  In a module to import,
it would include all the transitive things you've imported.  We could
make it type-dependent somehow, so e.g.  `export : [x: int, y: text]`
would only export the `x` and `y` in scope.  However, in that case we
might as well just write `[x=x, y=y]`, which is shorter; and we have
now added the shorthand syntax `[x,y]`.