WIP: get rid of recursion indicators in SDef and SLet
See discussion at #1148.  Adding an `open ... in ...` construct for
records means that we need to know type information in order to
calculate free variables.  Currently the only place we compute free
variables of an untyped term is in the parser, to mark whether `let`
and `def` are recursive.  However, we need that information when
capability checking (to see whether `CRecursion` capability is
needed), which currently happens independently of typechecking.