Simplify skippable list api (#1193)
The purpose of this refactoring is to decouple the key mapping from imperative list movement instructions.  In the new API (a function named `navigateList`), one supplies a declarative mapping of key events to a sum type that represents the desired movement.

As a client of the API, one's code reduces to this:
```
navigateList isSep e $ \case
    V.EvKey V.KUp [] -> Single Up
    V.EvKey V.KDown [] -> Single Down
    V.EvKey V.KHome [] -> Multi Up Max
    V.EvKey V.KEnd [] -> Multi Down Max
    V.EvKey V.KPageUp [] -> Multi Up Page
    V.EvKey V.KPageDown [] -> Multi Down Page
    _ -> None
```