finer texture for static (#1241)
Closes #1145.

Instead of using simple black or white cells to display static, use cells with any one of the 16 possible arrangements of on/off quarter-blocks.  This makes static appear at almost 2x resolution as before (though there is still a bit of blockiness since the quarter-blocks can't occur independently). See https://asciinema.org/a/582095 for a demonstration.